Activada cache en peticiones de listado
- Creada cache de 15 seg para los listados de registros

+++ app/Http/Controllers/FabricanteController.php
@@ -4,10 +4,12 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
 use Illuminate\Http\Request;
-
-// Cargamos Fabricante por que lo usamos mas abajo.
+use Illuminate\Support\Facades\Cache;
+// Cargamos Fabricante por que lo usamos más abajo.
 use App\Fabricante;
+
 use Response;
+
 class FabricanteController extends Controller {
 
 	/**
@@ -17,17 +19,23 @@ class FabricanteController extends Controller {
 	 */
 	public function index()
 	{
-		// Devolvemos un JSON con todos los fabricantes
-		//return Fabricante::all();
+		// return "En el index de Fabricante.";
+		// Devolvemos un JSON con todos los fabricantes.
+		// return Fabricante::all();
+		// Cache se actualizara con nuevos datos cada 15 segundos.
+		// 'cachefabricantes' es la clave con la que se almacenaran los registros
+		// obtenidos de Fabricante::all();
+		$fabricantes = Cache::remember('cachefabricantes',15/60,function(){
+			return Fabricante::all();
+		});
 
 
-		// Para devolver un JSON con codigo de respuesta HTTP.
-		return response()->json([
+		// Para devolver un JSON con código de respuesta HTTP SIN cache.
+		//return response()->json(['status'=>'ok', 'data'=>Fabricante::all()],200);	
 
-				'status'=>'ok',
-				'data' => Fabricante::all()
-			],200);
 
+		// Para devolver un JSON con codigo de respuesta HTTP CON cache.
+		return response()->json(['status'=>'ok', 'data'=>$fabricantes],200);
 	}
 
 	/**
@@ -35,14 +43,15 @@ class FabricanteController extends Controller {
 	 *
 	 * @return Response
 	 */
-	// No se utiliza este metodo por que se usaria para mostrar un formulario
-	// de creacion de Fabricantes. Y una API REST no hace eso.
+	// No se utiliza este método por que se usaría para mostrar un formulario
+	// de creación de Fabricantes. Y una API REST no hace eso.
 	/*
 	public function create()
 	{
 		//
 	}
 	*/
+
 	/**
 	 * Store a newly created resource in storage.
 	 *
@@ -50,22 +59,20 @@ class FabricanteController extends Controller {
 	 */
 	public function store(Request $request)
 	{
-		// Metodo llamado al hacer un POST.
+		// Método llamado al hacer un POST.
 		// Comprobamos que recibimos todos los campos.
-		if(!$request->input('nombre') || !$request->input('direccion')||!$request->input('telefono')){
-
-			//NO estamos recibiendo los campos necesarios. Devolvemos error.
-			return response()->json(['errors'=>Array(['code'=>422,'message'=>'Faltan datos necesarios para procesar el alta.'])],422);
+		if (!$request->input('nombre') || !$request->input('direccion') || !$request->input('telefono'))
+		{
+			// NO estamos recibiendo los campos necesarios. Devolvemos error.
+			return response()->json(['errors'=>array(['code'=>422,'message'=>'Faltan datos necesarios para procesar el alta.'])],422);
 		}
+
 		// Insertamos los datos recibidos en la tabla.
-		$nuevoFabricante = Fabricante::create($request->all());
+		$nuevoFabricante=Fabricante::create($request->all());
 
-		//Devolvemos la respuesta Http 201 (Created) + los datos del nuevo fabricante + una cabecera de Location
-
-		$respuesta = Response::make(json_encode(['data'=>$nuevoFabricante]),201)->header('Location','http://www.dominio.local/fabricantes/'.$nuevoFabricante->id)->header('Content-Type','application/json');
-
+		// Devolvemos la respuesta Http 201 (Created) + los datos del nuevo fabricante + una cabecera de Location + cabecera JSON
+		$respuesta= Response::make(json_encode(['data'=>$nuevoFabricante]),201)->header('Location','http://www.dominio.local/fabricantes/'.$nuevoFabricante->id)->header('Content-Type','application/json');
 		return $respuesta;
-
 	}
 
 	/**
@@ -76,25 +83,20 @@ class FabricanteController extends Controller {
 	 */
 	public function show($id)
 	{
-		//
-		/*
-		return response()->json([
+		// Corresponde con la ruta /fabricantes/{fabricante}
+		// Buscamos un fabricante por el ID.
+		$fabricante=Fabricante::find($id);
 
-				'status'=>'ok',
-				'data' => Fabricante::find($id)
-			],200);
-		*/
-		$fabricante = Fabricante::find($id);
-
-		if(!$fabricante){
-			// Se devuelve un array errors con los errores detectados y codigo 404
-			return response()->json([
-				'errors'=>Array(['code'=>404,'mensaje'=>'No se encuentra un fabricante con ese codigo.'])
-			],404);
+		// Chequeamos si encontró o no el fabricante
+		if (! $fabricante)
+		{
+			// Se devuelve un array errors con los errores detectados y código 404
+			return response()->json(['errors'=>Array(['code'=>404,'message'=>'No se encuentra un fabricante con ese código.'])],404);
 		}
 
-		// Devolvemos la informacion encontrada.
+		// Devolvemos la información encontrada.
 		return response()->json(['status'=>'ok','data'=>$fabricante],200);
+
 	}
 
 	/**
@@ -109,6 +111,7 @@ class FabricanteController extends Controller {
 		//
 	}
 	*/
+
 	/**
 	 * Update the specified resource in storage.
 	 *
@@ -117,60 +120,81 @@ class FabricanteController extends Controller {
 	 */
 	public function update($id,Request $request)
 	{
-		// Comprobamos si el fabricante existe. En otro caso devolvemos error
+		// Vamos a actualizar un fabricante.
+		// Comprobamos si el fabricante existe. En otro caso devolvemos error.
 		$fabricante=Fabricante::find($id);
 
 		// Si no existe mostramos error.
-		if(!$fabricante){
+		if (! $fabricante)
+		{
 			// Devolvemos error 404.
-			return response()->json(['errors'=>array(['code'=>404,'message'=>'No se encuentra un fabricante con ese codigo'])],404);
+			return response()->json(['errors'=>array(['code'=>404,'message'=>'No se encuentra un fabricante con ese código.'])],404);
 		}
-		$nombre =$request->input('nombre');
-		$direccion = $request->input('direccion');
+
+		// Almacenamos en variables para facilitar el uso, los campos recibidos.
+		$nombre=$request->input('nombre');
+		$direccion=$request->input('direccion');
 		$telefono=$request->input('telefono');
 
-		//Comprobamos si recibimos peticion PATH(parcial) o PUT(total)
-		if($request->method()=='PATCH'){
-
+		// Comprobamos si recibimos petición PATCH(parcial) o PUT (Total)
+		if ($request->method()=='PATCH')
+		{
 			$bandera=false;
-			//Actualizacion parcial de datos.
-			if($nombre != null && $nombre != ''){
+
+			// Actualización parcial de datos.
+			if ($nombre !=null && $nombre!='')
+			{
 				$fabricante->nombre=$nombre;
 				$bandera=true;
 			}
-			if($direccion != null && $direccion != ''){
+
+			// Actualización parcial de datos.
+			if ($direccion !=null && $direccion!='')
+			{
 				$fabricante->direccion=$direccion;
 				$bandera=true;
 			}
-			if($telefono != null && $telefono != ''){
+
+			// Actualización parcial de datos.
+			if ($telefono !=null && $telefono!='')
+			{
 				$fabricante->telefono=$telefono;
 				$bandera=true;
 			}
-			if($bandera){
 
-				//Grabamos el fabricante
+			if ($bandera)
+			{
+				// Grabamos el fabricante.
 				$fabricante->save();
-				//Devolvemos un codigo 200.
-				return response()->json(['status'=>'ok','data'=>$fabricante],200);
 
-			}else{
-				// Devolvemos un codigo 304 not modified.
-				return response()->json(['errors'=>array(['code'=>304,'message'=>'No se ha modificado ningun dato del fabricante'])],304);
+				// Devolvemos un código 200.
+				return response()->json(['status'=>'ok','data'=>$fabricante],200);
+			}
+			else
+			{
+				// Devolvemos un código 304 Not Modified.
+				return response()->json(['errors'=>array(['code'=>304,'message'=>'No se ha modificado ningún dato del fabricante.'])],304);
 			}
 		}
-		// Metodo PUT actualizamos todos los campos.
+
+
+		// Método PUT actualizamos todos los campos.
 		// Comprobamos que recibimos todos.
-		if(!$nombre || $direccion || $telefono){
-			// Se devuelve codigo 422 Unprocessable Entity.
+		if (!$nombre || !$direccion || !$telefono)
+		{
+			// Se devuelve código 422 Unprocessable Entity.
 			return response()->json(['errors'=>array(['code'=>422,'message'=>'Faltan valores para completar el procesamiento.'])],422);
-		}	
-		//Actualizamos los 3 campos.
+		}
+
+		// Actualizamos los 3 campos:
 		$fabricante->nombre=$nombre;
 		$fabricante->direccion=$direccion;
 		$fabricante->telefono=$telefono;
 
+		// Grabamos el fabricante
 		$fabricante->save();
 		return response()->json(['status'=>'ok','data'=>$fabricante],200);
+
 	}
 
 	/**
@@ -181,25 +205,41 @@ class FabricanteController extends Controller {
 	 */
 	public function destroy($id)
 	{
-		//Comprobamos si existe el registro indicado
-		$fabricante = Fabricante::find($id);
+		// Borrado de un fabricante.
+		// Ejemplo: /fabricantes/89 por DELETE
+		// Comprobamos si el fabricante existe o no.
+		$fabricante=Fabricante::find($id);
 
-		if(!$fabricante){
-			// Se devuelve un array errors con los errores detectados y codigo 404
-			return response()->json([
-				'errors'=>Array(['code'=>404,'mensaje'=>'No se encuentra un fabricante con ese codigo.'])
-			],404);
+		if (! $fabricante)
+		{
+			// Devolvemos error codigo http 404
+			return response()->json(['errors'=>array(['code'=>404,'message'=>'No se encuentra el fabricante con ese código.'])],404);
 		}
 
-		//Borramos el fabricante y devolvemos el codigo 204.
+		// Borramos el fabricante y devolvemos código 204
 		// 204 significa "No Content".
-		// Este codigo no muestra texto en el body.
-		// Si quisieramos ver el mensaje devolveriamos
-		// un 200.
+		// Este código no muestra texto en el body.
+		// Si quisiéramos ver el mensaje devolveríamos
+		// un código 200.
+		// Antes de borrarlo comprobamos si tiene aviones y si es así
+		// sacamos un mensaje de error.
+		// $aviones = $fabricante->aviones()->get();
+		$aviones = $fabricante->aviones;
+
+		if (sizeof($aviones) >0)
+		{
+			// Si quisiéramos borrar todos los aviones del fabricante sería:
+			// $fabricante->aviones->delete();
+
+			// Devolvemos un código 409 Conflict. 
+			return response()->json(['errors'=>array(['code'=>409,'message'=>'Este fabricante posee aviones y no puede ser eliminado.'])],409);
+		}
+
+		// Eliminamos el fabricante si no tiene aviones.
 		$fabricante->delete();
 
-		// Devolvemos la informacion encontrada.
-		return response()->json(['code'=>204,'message'=>'Se ha eliminado correctamente el fabricante'],204);
+		// Se devuelve código 204 No Content.
+		return response()->json(['code'=>204,'message'=>'Se ha eliminado correctamente el fabricante.'],204);
 	}
 
 }
